add condition for major release upgrade

+++ rocky-upgrade.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 
 # script by enigma158an201
-set -euo pipefail # set -euxo pipefail
+set -euxo pipefail # set -euxo pipefail
+
+#https://linux.how2shout.com/five-commands-to-check-the-almalinux-or-rocky-linux-version/
 
 ROCKY_VERSION=9.4-1.7
 REPO_URL="https://download.rockylinux.org/pub/rocky/9/BaseOS/x86_64/os/Packages/r"
@@ -27,9 +29,11 @@ main_rockylinux_upgrade() {
 		echo -e "\t>>> dnf found, this script will:\n 1. fetch updates\n 2. install updates\n 3. clean pkg archives\n 4.shutdown vm"
 	fi
 	update_dnf && clean_dnf #&& shutdown 0
-	sudo dnf install ${REPO_URL}/${RELEASE_PKG} ${REPO_URL}/${REPOS_PKG} ${REPO_URL}/${GPG_KEYS_PKG}
-	#restorecon -Rv /var/lib/rpm
-	#rpmdb --rebuilddb
+	if [[ "$(rpm -E "%{rhel}")" = "9" ]]; then 
+		sudo dnf install ${REPO_URL}/${RELEASE_PKG} ${REPO_URL}/${REPOS_PKG} ${REPO_URL}/${GPG_KEYS_PKG}
+		#restorecon -Rv /var/lib/rpm
+		#rpmdb --rebuilddb
+	fi
 }
 
 main_rockylinux_upgrade
